requirements.txt
flask
onnxruntime
numpy
opencv-python-headless
matplotlib
pillow
scipy
